+++ ios/Runner/AppDelegate.swift
@@ -2,8 +2,9 @@ import Flutter
 import UIKit
 import GoogleMaps
 
+// https://docs.flutter.dev/release/breaking-changes/uiscenedelegate
 @main
-@objc class AppDelegate: FlutterAppDelegate {
+@objc class AppDelegate: FlutterAppDelegate, FlutterImplicitEngineDelegate {
   override func application(
     _ application: UIApplication,
     didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
@@ -16,7 +17,11 @@ import GoogleMaps
       UNUserNotificationCenter.current().delegate = self as UNUserNotificationCenterDelegate
     }
     
-    GeneratedPluginRegistrant.register(with: self)
+    // GeneratedPluginRegistrant.register(with: self)
     return super.application(application, didFinishLaunchingWithOptions: launchOptions)
+  }
+    
+  func didInitializeImplicitFlutterEngine(_ engineBridge: FlutterImplicitEngineBridge) {
+    GeneratedPluginRegistrant.register(with: engineBridge.pluginRegistry)
   }
 }